+++ regions/shapes/polygon.py
@@ -62,6 +62,7 @@ class PolygonPixelRegion(PixelRegion):
 
     _params = ('vertices',)
     vertices = OneDPix('vertices')
+    mpl_artist = 'Patch'
 
     def __init__(self, vertices, meta=None, visual=None,
                  origin=PixCoord(0, 0)):
@@ -169,7 +170,7 @@ class PolygonPixelRegion(PixelRegion):
         xy = np.vstack([self.vertices.x - origin[0],
                         self.vertices.y - origin[1]]).transpose()
 
-        mpl_kwargs = self._define_mpl_kwargs(artist='Patch')
+        mpl_kwargs = self.visual.define_mpl_kwargs(self.mpl_artist)
         mpl_kwargs.update(kwargs)
 
         return Polygon(xy=xy, **mpl_kwargs)